♻️ refactor(cli): extract spinner, validation, and contracts
