+++ src/Nupeek.Cli/Contracts/InlineSourceResult.cs
@@ -1,0 +1,7 @@
+namespace Nupeek.Cli;
+
+internal sealed record InlineSourceResult(
+    string? Content,
+    int? MaxChars,
+    int? OriginalChars,
+    bool Truncated);
